% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eddies_matching.r
\name{eddies_edges}
\alias{eddies_edges}
\title{Compute the coordinates of the edges of an eddy}
\usage{
eddies_edges(x, ...)
}
\arguments{
\item{x}{An "eddy" object as obtained using \code{\link{eddies}}.}

\item{...}{arguments to be passed to \code{\link{ell.coords}}. Setting \code{n}
to a lower level speeds up computing time.}
}
\description{
Compute the coordinates of the edges of an eddy
}
\keyword{internal}

