% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg_tools.r
\name{ellipses}
\alias{ellipses}
\title{Plot ellipses on a existing graph}
\usage{
ellipses(x0, y0, major, minor, agl, n = 50, ...)
}
\arguments{
\item{x0}{x coordinate of the center}

\item{y0}{y coordinate of the center}

\item{major}{semi-length of the major axe}

\item{minor}{semi-length of the minor axe}

\item{agl}{rotation angle  from the x-axis}

\item{n}{the number of points to use}

\item{...}{other arguments to be passed to \code{\link{lines}}}
}
\description{
Plot ellipses on a existing graph
}
\examples{
plot(0:1, 0:1)
ellipses(.5, .5, .3, .1, pi/6)

# Arguments are recycled
plot(0:1, 0:1)
ellipses(c(.5, .75), c(.5, .75), .3, .1, c(pi/6, -pi/2))
}
\seealso{
\code{\link{ell.coords}}
}
\keyword{internal}

