% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eddies_matching.r
\name{match_loc2eddies}
\alias{match_loc2eddies}
\title{Match a location to the edges of an eddy}
\usage{
match_loc2eddies(x, lon, lat, edges_coords = NULL, ...)
}
\arguments{
\item{x}{An "eddy" object as obtained using \code{\link{eddies}}.}

\item{lon}{location longitude}

\item{lat}{location latitude}

\item{edges_coords}{the coordinates of the edges of the eddies in \code{x}.
This arguments allows to speed up calculation when this function is called
from \code{\link{match_track2eddies}} and the same eddies are considered again
and again. See \code{\link{eddies_edges}}.}

\item{...}{arguments to be passed to \code{\link{ell.coords}}. Setting \code{n}
to a lower level speeds up computing time.}
}
\description{
Match a location to the edges of an eddy
}
\seealso{
\code{\link{match_track2eddies}}
}

