% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{read_eddy}
\alias{read_eddy}
\title{Read an eddy file}
\usage{
read_eddy(x, use.rds = TRUE, rds.save = TRUE, overwrite = FALSE)
}
\arguments{
\item{x}{filename}

\item{use.rds}{Should the \code{".rds"} copy be used to speed up the function
if it exists ? Otherwise, the original \code{".mat"} (matlab) file will be used.}

\item{rds.save}{Should a \code{".rds"} copy of \code{x} be saved for a faster
access  to data later on ?}

\item{overwrite}{Should the \code{"rds"} file be overwritten if it already exists ?}
}
\description{
Read an eddy file
}
\details{
The function uses the path provided by \code{options("eddies.area")}
to subset a geographical region of focus. This behavior can be turned off by
setting this option to \code{NULL}
}
\examples{
\dontrun{
edd <- read_eddy("F:/data/eddies/2010/anticyc_20100101.mat")
}
}

