% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{select_eddy}
\alias{select_eddy}
\title{Find eddy files}
\usage{
select_eddy(date, type = c("both", "cyc", "antcyc"), ext = c("mat", "rds",
  "any"))
}
\arguments{
\item{date}{Date of the eddy}

\item{type}{Type of the eddy}
}
\description{
Find eddy files
}
\details{
The function uses the path provided by \code{options("eddies.path")}
to access to the eddies database. The database only includes years from
2010 to 2014 but years from 1993 to 2009 are also available from
the downloaded database
(\url{http://datadryad.org/resource/doi:10.5061/dryad.gp40h/1}). To get
more recent data in a similar format
the software used in the article (matlab) was made available
online at \url{https://github.com/jfaghm/OceanEddies}.
}
\examples{
\dontrun{
days <- as.POSIXct(c("2010-01-01", "2011-01-01"))
select_eddy(day, "both")
}
}

